Moving files to "almost" recomended places. RTThreadBuffer still needs a better one.
