--- alot/foreign/retrothread/__init__.py
@@ -1,1 +1,0 @@
-from retrothread import RTThreadBuffer
